backend/requirements.txt
fastapi
uvicorn[standard]
celery[redis]
sqlalchemy
psycopg2-binary      # PostgreSQL を使う場合
python-dotenv